+++ Assets/00_WorkSpace/PJW/Scirpts/Player/DollyTrackPlayerSpawner.cs
@@ -1,40 +1,52 @@
-using Cinemachine;
-using Photon.Pun;
-using Photon.Realtime;
 using System;
-using System.Linq;
+using Photon.Pun;
 using UnityEngine;
+using Cinemachine;
+
 namespace PJW
 {
-    public class PlayerSpawner : MonoBehaviourPunCallbacks
+    public class DollyTrackPlayerSpawner : MonoBehaviourPunCallbacks
     {
         [Header("Tracks / Prefab")]
         [SerializeField] private CinemachinePathBase[] tracks;
         [SerializeField] private string playerPrefabName = "Player";
 
         [Header("Placement")]
-        [SerializeField] private float heightOffset = 0.5f;
-        [SerializeField] private float spawnT = 0f;
-        [SerializeField] private bool faceAlongTrack = true; private bool hasSpawned;
-        public override void OnJoinedRoom() { TrySpawn(); }
+        private float spawnT = 0f; // 트랙의 어느 지점에서 스폰할지
+        private bool faceAlongTrack = true; // 트랙 진행 방향을 바라보게 스폰할지 여부
+
+        private bool hasSpawned;
+
+        private void Start()
+        {
+            if (PhotonNetwork.InRoom && !hasSpawned)
+            {
+                TrySpawn();
+            }
+        }
+
         private void TrySpawn()
         {
             if (hasSpawned) return;
 
-            // 플레이어 순서: ActorNumber 기준으로 정렬
-            Player[] sorted = PhotonNetwork.PlayerList.OrderBy(p => p.ActorNumber).ToArray();
+            int wantedIndex = PhotonNetwork.LocalPlayer.ActorNumber - 1;
+            int trackIndex = Mathf.Clamp(wantedIndex, 0, tracks.Length - 1);
+            var myTrack = tracks[trackIndex];
 
-            int myIndex = Array.FindIndex(sorted, p => p == PhotonNetwork.LocalPlayer);
-            if (myIndex < 0 || myIndex >= tracks.Length) { return; }
+            if (myTrack == null)
+            {
+                return;
+            }
 
-            // 내 차례에 해당하는 트랙 선택
-            CinemachinePathBase myTrack = tracks[myIndex];
-            if (myTrack == null) { return; }
-            Vector3 pos = myTrack.EvaluatePositionAtUnit(spawnT, CinemachinePathBase.PositionUnits.Normalized); 
-            pos += Vector3.up * heightOffset; 
-            Quaternion rot = Quaternion.identity; 
-            if (faceAlongTrack) { rot = myTrack.EvaluateOrientationAtUnit(spawnT, CinemachinePathBase.PositionUnits.Normalized); }
-            PhotonNetwork.Instantiate(playerPrefabName, pos, rot); hasSpawned = true;
+            Vector3 pos = myTrack.EvaluatePositionAtUnit(spawnT, CinemachinePathBase.PositionUnits.Normalized) + Vector3.up; // 스폰을 0에서 시작하게 해주는 기능
+            Quaternion rot = Quaternion.identity;
+
+            // 프리팹 존재 검사
+            var prefab = Resources.Load<GameObject>(playerPrefabName);
+            if (prefab == null) return;
+            var go = PhotonNetwork.Instantiate(playerPrefabName, pos, rot);
+
+            hasSpawned = true;
         }
     }
 }
