--- Assets/00_WorkSpace/PJW/Scirpts/Track/DollyTrackPlayerSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e75b493fe26a21c4d9a6c610d2c97877
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
