+++ Assets/00_WorkSpace/PJW/Scirpts/Track/TrackRegistry.cs
@@ -7,7 +7,6 @@ namespace PJW
     {
         public static TrackRegistry Instance { get; private set; }
 
-        [Header("전체 트랙 리스트")]
         public CinemachinePathBase[] tracks;
 
         private void Awake()
@@ -21,17 +20,9 @@ namespace PJW
             Instance = this;
         }
 
-        /// <summary>
-        /// 로컬 플레이어의 ActorNumber에 해당하는 트랙 반환
-        /// </summary>
         public CinemachinePathBase GetPathForLocalPlayer(int actorNumber)
         {
-            if (tracks == null || tracks.Length == 0)
-                return null;
-
             int index = actorNumber - 1;
-            if (index < 0 || index >= tracks.Length)
-                return null;
 
             return tracks[index];
         }
